Moved code; added comment #3926
